requirements.txt
beautifulsoup4==4.13.3
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.43.2
